requirements.txt
flask
supabase
openpyxl
python-dotenv